LU-5577 obdclass: cleanup comparison between signed and unsigned
Change arguments for page offset from int to loff_t.
Change type of counter and size variables to size_t.

+++ lustre/mdd/mdd_object.c
@@ -1861,7 +1861,7 @@ static int mdd_readpage_sanity_check(const struct lu_env *env,
 }
 
 static int mdd_dir_page_build(const struct lu_env *env, union lu_page *lp,
-			      int nob, const struct dt_it_ops *iops,
+			      size_t nob, const struct dt_it_ops *iops,
 			      struct dt_it *it, __u32 attr, void *arg)
 {
 	struct lu_dirpage	*dp = &lp->lp_dir;
@@ -1879,8 +1879,8 @@ static int mdd_dir_page_build(const struct lu_env *env, union lu_page *lp,
 
         ent  = area;
         do {
-                int    len;
-                int    recsize;
+		int    len;
+		size_t recsize;
 
                 len = iops->key_size(env, it);
